+++ client/state/selectors/get-editor-url.js
@@ -1,10 +1,12 @@
 /**
  * Internal dependencies
  */
-import getGutenbergEditorUrl from 'calypso/state/selectors/get-gutenberg-editor-url';
+import { isEligibleForGutenframe } from 'calypso/state/gutenberg-iframe-eligible/is-eligible-for-gutenframe';
+import { getSelectedEditor } from 'calypso/state/selectors/get-selected-editor';
+import { getSiteAdminUrl, getSiteSlug } from 'calypso/state/sites/selectors';
+import { getEditorPath } from 'calypso/state/editor/selectors';
+import { addQueryArgs } from 'calypso/lib/route';
 import isClassicEditorForced from 'calypso/state/selectors/is-classic-editor-forced';
-import isEligibleForGutenframe from 'calypso/state/gutenberg-iframe-eligible/is-eligible-for-gutenframe';
-import { getSiteAdminUrl } from 'calypso/state/sites/selectors';
 
 export const getEditorUrl = ( state, siteId, postId = null, postType = 'post' ) => {
 	// isEligibleForGutenframe deals with the server side checks for classic editor plugin etc.
@@ -17,10 +19,23 @@ export const getEditorUrl = ( state, siteId, postId = null, postType = 'post' ) 
 			url = `${ siteAdminUrl }post.php?post=${ postId }&action=edit`;
 		}
 
+		if ( 'gutenberg-redirect-and-style' === getSelectedEditor( state, siteId ) ) {
+			url = addQueryArgs( { calypsoify: '1' }, url );
+		}
+
 		return url;
 	}
 
-	return getGutenbergEditorUrl( state, siteId, postId, postType );
+	if ( postId ) {
+		return getEditorPath( state, siteId, postId, postType );
+	}
+
+	const siteSlug = getSiteSlug( state, siteId );
+
+	if ( 'post' === postType || 'page' === postType ) {
+		return `/${ postType }/${ siteSlug }`;
+	}
+	return `/edit/${ postType }/${ siteSlug }`;
 };
 
 export default getEditorUrl;
